+++ NEWS.md
@@ -5,6 +5,7 @@
 API:
  * DBus::Object#object_server replaces @service (which still works) and the short-lived
    @connection
+ * ObjectServer#export will raise if the path is already taken by an object
  * ObjectServer#unexport now also accepts an object path
  * Connection#object_server can export objects even without requesting any
    service name ([#49][], in beta1 already).